+++ src/ResourceManager/Compute/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Current Release
+* `New-AzureRmVm` and `New-AzureRmVmss` (simple parameter set) support assigning user defined and(or) system defined identities to the VM(s).
 * VMSS Redeploy and PerformMaintenance feature
     -  Add new switch parameter -Redeploy and -PerformMaintenance to `Set-AzureRmVmss` and `Set-AzureRmVmssVM`
 * Add DisableVMAgent switch parameter to `Set-AzureRmVMOperatingSystem` cmdlet